requirements.txt
Django==3.0.6
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.10.0
graphql-core==2.3.1
graphql-relay==2.0.1
